Upgrade UICurrency.
